% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_khat.R
\name{calculate.khat}
\alias{calculate.khat}
\title{Calculate K hat}
\usage{
calculate.khat(pro_mat, qc, max_k)
}
\arguments{
\item{pro_mat}{A N by N matrix of pairwise proximities.}

\item{qc}{A string specifying the cluster quality criteria.  Choices are "ASW", "CH" and "Gap", for average silhouette width, Calinski criteria and Gap statistic.}

\item{max_k}{An integer for the largest K.}
}
\value{
An integer for the estimated number of clusters.
}
\description{
Calculate K hat
}
\examples{
set.seed(808)
N <- 50
n <- 50
dat <- data.frame(id = rep(1:N, each = n),
time = rep(seq(0, 1, length.out = n), times = N),
y = rnorm(25))
cp <- create.pred(dat, 3)
pro_mat <- create.pro_mat(cp, "ED")
khat <- calculate.khat(pro_mat, "CH", 6)
}
