% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{create.data}
\alias{create.data}
\title{Create data to be used in simulation}
\usage{
create.data(sp = "regular", Ng = "50:50:50", ni = 50, sigma = 1)
}
\arguments{
\item{sp}{Spacing between a subject's observations.  Options include, \code{regular}, \verb{partially random} and \verb{totally random}.}

\item{Ng}{A string with the number of subjects in each group, separated by colons (:).}

\item{ni}{Number of observations per subject.}

\item{sigma}{Standard deviation of the normal error term.}
}
\value{
A \code{list} with data in long and wide form, true cluster assignments and true number of clusters.
}
\description{
Create data to be used in simulation
}
\examples{
set.seed(808)
cd <- create.data(sigma = 0.1)
plot(response ~ time, data = cd$dat, type = "p")
}
