% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred.R
\name{create.pred}
\alias{create.pred}
\title{Create matix of predictions}
\usage{
create.pred(dat, lo)
}
\arguments{
\item{dat}{\code{data.frame} with columns \code{id}, \code{time}, \code{y}.}

\item{lo}{number of predicted points.}
}
\value{
\code{list}.
}
\description{
Create matix of predictions
}
\examples{
set.seed(808)
N <- 50
n <- 50
dat <- data.frame(id = rep(1:N, each = n),
time = rep(seq(0, 1, length.out = n), times = N),
y = rnorm(25))
cp <- create.pred(dat, 3)
}
