% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pro_mat.R
\name{create.pro_mat}
\alias{create.pro_mat}
\title{Create prediction matrix}
\usage{
create.pro_mat(co, vc)
}
\arguments{
\item{co}{A list from \code{create.pred}.}

\item{vc}{A string specifying the proximity.  Choices are "II", "ED" and "MD" for individual information, euclidean distance and mahalanobis distance.}
}
\value{
A N by N matrix of pairwise proximity.
}
\description{
Create prediction matrix
}
\examples{
set.seed(808)
N <- 50
n <- 50
dat <- data.frame(id = rep(1:N, each = n),
time = rep(seq(0, 1, length.out = n), times = N),
y = rnorm(25))
cp <- create.pred(dat, 3)
pro_mat <- create.pro_mat(cp, "ED")
}
