% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_methods.R
\name{iteration}
\alias{iteration}
\title{An iteration of a simulation run}
\usage{
iteration()
}
\value{
A \code{data.frame}.
}
\description{
An iteration of a simulation run
}
\examples{
set.seed(808)
iteration()
}
