% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic2gam.R
\name{magic2gam}
\alias{magic2gam}
\title{Update \code{mgcv::gam} with results from \code{mgcv::magic}}
\usage{
magic2gam(gam_F, magic, w)
}
\arguments{
\item{gam_F}{Output of \code{mgcv::gam}.}

\item{magic}{Output of \code{mgcv::magic}.}

\item{w}{Variance-covariance matrix that necessitated the use of \code{mgcv::magic}.}
}
\value{
A \code{mgcv::gamObject}.
}
\description{
Update \code{mgcv::gam} with results from \code{mgcv::magic}
}
\examples{
# to do.
}
