% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmkl.R
\name{pmkl}
\alias{pmkl}
\title{Prediction model-based k-medoids for longitudinal data}
\usage{
pmkl(dat, lo = 10, vc = "II", qc = "CH", max_k = 6, clusGap_boot = 100)
}
\arguments{
\item{dat}{A \code{data.frame} of longitudinal data in long form  with columns \code{id}, \code{time}, \code{response}.}

\item{lo}{Integer for the number of points to predict at.}

\item{vc}{A string specifying the proximity.  Choices are, \code{II} (individual information), \code{ED} (Euclidean distance) and \code{MD} (Mahalanobis distance).}

\item{qc}{The quality criterion used to estimate the number of clusters.  Choices are \code{CH} (Calinski-Harabasz), \code{ASW} (Average Silhouette Width) and \code{Gap} (Gap statistic).}

\item{max_k}{The maximum number of clusters to consider.  The method considers 2:max_k.}

\item{clusGap_boot}{Argument passed to \code{cluster::clusGap} for the number of boot strap iterations.}
}
\value{
The results from \code{cluster::pam}; a \code{pam} object.
}
\description{
Prediction model-based k-medoids for longitudinal data
}
\examples{
set.seed(808)
N <- 50
n <- 50
dat <- data.frame(id = rep(1:N, each = n),
time = rep(seq(0, 1, length.out = n), times = N),
response = rnorm(25))
mod <- pmkl(dat)
}
