% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_methods.R
\name{iteration}
\alias{iteration}
\title{An iteration of a simulation run}
\usage{
iteration()
}
\value{
A \code{data.frame}.
}
\description{
An iteration of a simulation run
}
\examples{
\dontrun{
set.seed(808)
sp <- "regular"
Ng <- "50:50:50"
ni <- 50
sigma <- 1
iteration()}
}
